delete alert

+++ src/app/admin components/accounts-management/accounts-management.component.ts
@@ -1,13 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { UserManagementService } from '../services/user management service/user-management.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-accounts-management',
   templateUrl: './accounts-management.component.html',
-  styleUrl: './accounts-management.component.css'
+  styleUrls: ['./accounts-management.component.css'] // Fixed 'styleUrls' instead of 'styleUrl'
 })
 export class AccountsManagementComponent implements OnInit {
-  users!: any[]; 
+  users!: any[];
 
   constructor(private userManagementService: UserManagementService) {}
 
@@ -18,11 +19,25 @@ export class AccountsManagementComponent implements OnInit {
   }
 
   toggleUserStatus(user: any) {
-    const newStatus = user.status === 'active' ? false : true;
-    this.userManagementService.updateUserStatus(user.uid, newStatus);
+    const newStatus = user.status === 'active' ? 'inactive' : 'active';
+    this.userManagementService.updateUserStatus(user.uid, newStatus === 'active');
   }
 
   deleteUser(userId: string) {
-    this.userManagementService.deleteUser(userId);
+    Swal.fire({
+      title: 'Are you sure you want to delete this user?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.userManagementService.deleteUser(userId);
+      }
+    });
   }
+  
+  
 }
